// alpha-backend-api-gateway-v1/src/interfaces/response.interface.ts
export interface IErrorResponse {
    statusCode: number;
    error: {
        code: number;
        urlPath: string;
        timestamp: string;
        message: string;
        details?: Array<{
            field: string;
            message: string;
        }>;
        debug?: {
            stack: string;
        };
    };
}